requirements.txt
deta==1.1.0
python-dotenv==0.21.1
rich==13.3.1
